+++ src/main/java/com/github/cowwoc/docker/internal/util/ClientRequests.java
@@ -191,7 +191,7 @@ public final class ClientRequests
 	 * @param response the server response
 	 * @return the response HTTP version, status code and reason phrase
 	 */
-	public String getStatusLine(ContentResponse response)
+	public String getStatusLine(Response response)
 	{
 		String reason = response.getReason();
 		if (reason == null)
@@ -206,7 +206,7 @@ public final class ClientRequests
 	 * @param response the server response
 	 * @return the string representation of the response's headers
 	 */
-	private String getHeadersAsString(ContentResponse response)
+	private String getHeadersAsString(Response response)
 	{
 		StringJoiner result = new StringJoiner("\n");
 		for (HttpField header : response.getHeaders())
@@ -218,17 +218,20 @@ public final class ClientRequests
 	 * @param response the server response
 	 * @return the {@code String} representation of the response
 	 */
-	public String toString(ContentResponse response)
+	public String toString(Response response)
 	{
 		requireThat(response, "response").isNotNull();
 		StringJoiner resultAsString = new StringJoiner("\n");
 		resultAsString.add(getStatusLine(response)).
 			add(getHeadersAsString(response));
-		String responseBody = response.getContentAsString();
-		if (!responseBody.isEmpty())
+		if (response instanceof ContentResponse contentResponse)
 		{
-			resultAsString.add("");
-			resultAsString.add(responseBody);
+			String responseBody = contentResponse.getContentAsString();
+			if (!responseBody.isEmpty())
+			{
+				resultAsString.add("");
+				resultAsString.add(responseBody);
+			}
 		}
 		return resultAsString.toString();
 	}